requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorlog==6.8.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
django-admin-tools==0.9.3
django-allauth==0.62.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-jazzmin==3.0.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
hiredis==2.3.2
httplib2==0.22.0
idna==3.7
inflection==0.5.1
oauthlib==3.2.2
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==1.7.1
pyparsing==3.1.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
random-password-generator==2.2.0
redis==5.0.4
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
social-auth-app-django==5.4.1
social-auth-core==4.5.4
soupsieve==2.5
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1
uuid==1.30
uuid64==0.2.0
